Change from UserControl to Custom Control.
