--- src/WpfMath/Controls/FormulaControl.xaml
@@ -1,33 +1,0 @@
-<UserControl x:Class="WpfMath.Controls.FormulaControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:WpfMath.Controls"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-    <UserControl.Resources>
-        <ControlTemplate x:Key="DefaultErrorTemplate">
-            <DockPanel>
-                <Border BorderBrush="Red" BorderThickness="1">
-                    <AdornedElementPlaceholder/>
-                </Border>
-            </DockPanel>
-        </ControlTemplate>
-    </UserControl.Resources>
-    
-    <UserControl.Style>
-        <Style TargetType="UserControl">
-            <Setter Property="local:FormulaControl.ErrorTemplate" Value="{StaticResource DefaultErrorTemplate}"/>
-            <Style.Triggers>
-                <DataTrigger Binding="{Binding HasError, RelativeSource={RelativeSource Self}}" Value="True">
-                    <Setter Property="Template" Value="{Binding ErrorTemplate, RelativeSource={RelativeSource Self}}"/>
-                </DataTrigger>
-            </Style.Triggers>
-        </Style>
-    </UserControl.Style>
-    
-    <Grid>
-        <local:VisualContainerElement x:Name="formulaContainerElement"/>
-    </Grid>
-</UserControl>
